+++ photon-model-tasks/src/test/java/com/vmware/photon/controller/model/tasks/TaskServices.java
@@ -32,6 +32,7 @@ public class TaskServices {
                 ResourceAllocationTaskService.createFactory(),
                 ResourceEnumerationTaskService.createFactory(),
                 ResourceRemovalTaskService.createFactory(),
-                SnapshotTaskService.createFactory());
+                SnapshotTaskService.createFactory(),
+                ScheduledTaskService.createFactory());
     }
 }